What is the next step in Conway's Life?
Given a matrix A that represents the state of Conway's game of Life at one instant (time t=n), return the matrix B that represents the state of the game at the next instant (time t=n+1).
Assume a toroidal game board, so the edges of the matrix wrap left to right and top to bottom.
Examples:
 Input  A = [ 1   1   0   0
              0   1   0   0
              1   1   0   0
              0   0   0   0 ]

 Output B = [ 1   1   0   0
              0   0   1   0
              1   1   0   0
              0   0   0   0 ]

 Input  A = [ 0   1   1   0
              1   1   1   0
              0   0   1   0
              0   0   0   0 ]

 Output B = [ 1   0   1   1
              1   0   0   0
              0   0   1   1
              0   1   1   0 ]